+++ src/templates/movie-card.hbs
@@ -13,7 +13,7 @@
       </div>
       <div class='movie__text'>
         <h2 class='movie__title'>{{original_title}}</h2>
-        <p class='movie__genre'>{{genreName}} | {{release_date}}</p>
+        <p class='movie__genre'>{{genre_ids}} | {{release_date}}</p>
       </div>
       <div class='one-card_overlay'>
         <div class='one-card__rate'>{{vote_average}}</div>
